+++ Interface/Addons/WeakAuras/Prototypes.lua
@@ -261,96 +261,96 @@ function Private.InitializeEncounterAndZoneLists()
       {
         L["Karazhan"],
         {
-          { L["Attumen the Huntsman"], 2444 },
-          { L["Moroes"], 2445 },
-          { L["Maiden of Virtue"], 2446 },
-          { L["Opera Hall"], 2447 },
-          { L["The Curator"], 2448 },
-          { L["Terestian Illhoof"], 2449 },
-          { L["Shade of Aran"], 2450 },
-          { L["Netherspite"], 2451 },
-          { L["Chess Event"], 2452 },
-          { L["Prince Malchezaar"], 2453 },
-          { L["Nightbane"], 2454 },
+          { L["Attumen the Huntsman"], 652 },
+          { L["Moroes"], 653 },
+          { L["Maiden of Virtue"], 654 },
+          { L["Opera Hall"], 655 },
+          { L["The Curator"], 656 },
+          { L["Terestian Illhoof"], 657 },
+          { L["Shade of Aran"], 658 },
+          { L["Netherspite"], 659 },
+          { L["Chess Event"], 660 },
+          { L["Prince Malchezaar"], 661 },
+          { L["Nightbane"], 662 },
         }
       },
       {
         L["Gruul's Lair"],
         {
-          { L["High King Maulgar"], 2455 },
-          { L["Gruul the Dragonkiller"], 2456 },
+          { L["High King Maulgar"], 649 },
+          { L["Gruul the Dragonkiller"], 650 },
         }
       },
       {
         L["Magtheridon's Lair"],
         {
-          { L["Magtheridon"], 2457 },
+          { L["Magtheridon"], 651 },
         }
       },
       {
         L["Coilfang: Serpentshrine Cavern"],
         {
-          { L["Hydross the Unstable"], 2458 },
-          { L["The Lurker Below"], 2459 },
-          { L["Leotheras the Blind"], 2460 },
-          { L["Fathom-Lord Karathress"], 2461 },
-          { L["Morogrim Tidewalker"], 2462 },
-          { L["Lady Vashj"], 2463 },
+          { L["Hydross the Unstable"], 623 },
+          { L["The Lurker Below"], 624 },
+          { L["Leotheras the Blind"], 625 },
+          { L["Fathom-Lord Karathress"], 626 },
+          { L["Morogrim Tidewalker"], 627 },
+          { L["Lady Vashj"], 628 },
         }
       },
       {
         L["Tempest Keep"],
         {
-          { L["Al'ar"], 2464 },
-          { L["Void Reaver"], 2465 },
-          { L["High Astromancer Solarian"], 2466 },
-          { L["Kael'thas Sunstrider"], 2467 },
+          { L["Al'ar"], 730 },
+          { L["Void Reaver"], 731 },
+          { L["High Astromancer Solarian"], 732 },
+          { L["Kael'thas Sunstrider"], 733 },
         }
       },
       {
         L["The Battle for Mount Hyjal"],
         {
-          { L["Rage Winterchill"], 2468 },
-          { L["Anetheron"], 2469 },
-          { L["Kaz'rogal"], 2470 },
-          { L["Azgalor"], 2471 },
-          { L["Archimonde"], 2472 },
+          { L["Rage Winterchill"], 618 },
+          { L["Anetheron"], 619 },
+          { L["Kaz'rogal"], 620 },
+          { L["Azgalor"], 621 },
+          { L["Archimonde"], 622 },
         }
       },
       {
         L["Black Temple"],
         {
-          { L["High Warlord Naj'entus"], 2473 },
-          { L["Supremus"], 2474 },
-          { L["Shade of Akama"], 2475 },
-          { L["Teron Gorefiend"], 2476 },
-          { L["Gurtogg Bloodboil"], 2477 },
-          { L["Reliquary of Souls"], 2478 },
-          { L["Mother Shahraz"], 2479 },
-          { L["The Illidari Council"], 2480 },
-          { L["Illidan Stormrage"], 2481 },
+          { L["High Warlord Naj'entus"], 601 },
+          { L["Supremus"], 602 },
+          { L["Shade of Akama"], 603 },
+          { L["Teron Gorefiend"], 604 },
+          { L["Gurtogg Bloodboil"], 605 },
+          { L["Reliquary of Souls"], 606 },
+          { L["Mother Shahraz"], 607 },
+          { L["The Illidari Council"], 608 },
+          { L["Illidan Stormrage"], 609 },
         }
       },
       {
         L["Zul'Aman"],
         {
-          { L["Akil'zon"], 2482 },
-          { L["Nalorakk"], 2483 },
-          { L["Jan'alai"], 2484 },
-          { L["Halazzi"], 2485 },
-          { L["Hex Lord Malacrass"], 2486 },
-          { L["Daakara"], 2487 },
+          { L["Akil'zon"], 1189 },
+          { L["Nalorakk"], 1190 },
+          { L["Jan'alai"], 1191 },
+          { L["Halazzi"], 1192 },
+          { L["Hex Lord Malacrass"], 1193 },
+          { L["Daakara"], 1194 },
         }
       },
       {
         L["The Sunwell Plateau"],
         {
-          { L["Kalecgos"], 2488 },
-          { L["Brutallus"], 2489 },
-          { L["Felmyst"], 2490 },
-          { L["Eredar Twins"], 2491 },
-          { L["M'uru"], 2492 },
-          { L["Kil'jaeden"], 2493 },
+          { L["Kalecgos"], 724 },
+          { L["Brutallus"], 725 },
+          { L["Felmyst"], 726 },
+          { L["Eredar Twins"], 727 },
+          { L["M'uru"], 728 },
+          { L["Kil'jaeden"], 729 },
         }
       },
     }
@@ -416,7 +416,7 @@ local function get_encounters_list()
 end
 
 local function get_zoneId_list()
-  if not WeakAuras.IsRetail() then return "" end
+  if WeakAuras.IsClassic() then return "" end
   local currentmap_id = C_Map.GetBestMapForUnit("player")
   local currentmap_info = C_Map.GetMapInfo(currentmap_id)
   local currentmap_name = currentmap_info and currentmap_info.name or ""
@@ -1573,21 +1573,21 @@ Private.load_prototype = {
       hidden = true,
       init = "arg",
       test = "true",
-      enable = WeakAuras.IsRetail(),
+      enable = not WeakAuras.IsClassic(),
     },
     {
       name = "zonegroupId",
       hidden = true,
       init = "arg",
       test = "true",
-      enable = WeakAuras.IsRetail(),
+      enable = not WeakAuras.IsClassic(),
     },
     {
       name = "zoneIds",
       display = L["Zone ID(s)"],
       type = "string",
-      enable = WeakAuras.IsRetail(),
-      hidden = not WeakAuras.IsRetail(),
+      enable = not WeakAuras.IsClassic(),
+      hidden = WeakAuras.IsClassic(),
       events = {"ZONE_CHANGED", "ZONE_CHANGED_INDOORS", "ZONE_CHANGED_NEW_AREA", "VEHICLE_UPDATE"},
       desc = get_zoneId_list,
       preamble = "local zoneChecker = WeakAuras.ParseZoneCheck(%q)",
@@ -2394,6 +2394,14 @@ Private.event_prototypes = {
         conditionType = "number"
       },
       {
+        name = "deficit",
+        display = WeakAuras.newFeatureString .. L["Health Deficit"],
+        type = "number",
+        init = "total - value",
+        store = true,
+        conditionType = "number"
+      },
+      {
         name = "showAbsorb",
         display = L["Show Absorb"],
         type = "toggle",
@@ -2879,6 +2887,14 @@ Private.event_prototypes = {
         display = L["Power (%)"],
         type = "number",
         init = "total ~= 0 and (value / total) * 100 or nil",
+        store = true,
+        conditionType = "number"
+      },
+      {
+        name = "deficit",
+        display = WeakAuras.newFeatureString .. L["Power Deficit"],
+        type = "number",
+        init = "total - value",
         store = true,
         conditionType = "number"
       },
@@ -7692,7 +7708,7 @@ Private.event_prototypes = {
         "PLAYER_TARGET_CHANGED"
       },
       ["unit_events"] = {
-        ["player"] = {"UNIT_STATS", "UNIT_ATTACK_POWER", "UNIT_AURA"}
+        ["player"] = {"UNIT_STATS", "UNIT_ATTACK_POWER", "UNIT_AURA", "PLAYER_DAMAGE_DONE_MODS"}
       }
     },
     internal_events = function(trigger, untrigger)
@@ -7785,6 +7801,26 @@ Private.event_prototypes = {
         init = "WeakAuras.GetCritChance()",
         store = true,
         conditionType = "number"
+      },
+      {
+        name = "hitrating",
+        display = L["Hit Rating"],
+        type = "number",
+        init = "max(GetCombatRating(CR_HIT_MELEE), GetCombatRating(CR_HIT_RANGED), GetCombatRating(CR_HIT_SPELL))",
+        store = true,
+        enable = WeakAuras.IsBCC(),
+        conditionType = "number",
+        hidden = not WeakAuras.IsBCC()
+      },
+      {
+        name = "hitpercent",
+        display = L["Hit (%)"],
+        type = "number",
+        init = "WeakAuras.GetHitChance()",
+        store = true,
+        conditionType = "number",
+        enable = WeakAuras.IsBCC(),
+        hidden = not WeakAuras.IsBCC()
       },
       {
         name = "hasterating",