+++ Interface/Addons/WeakAuras/RegionTypes/Model.lua
@@ -80,6 +80,7 @@ local regionFunctions = {
 local function create(parent)
   -- Main region
   local region = CreateFrame("FRAME", nil, UIParent);
+  region.regionType = "model"
   region:SetMovable(true);
   region:SetResizable(true);
   region:SetMinResize(1, 1);
@@ -161,7 +162,7 @@ local function ConfigureModel(region, model, data)
   end
 
   -- Enable model animation
-  if(data.advance) then
+  if(data.advance and model:HasAnimation(data.sequence)) then
     model:SetAnimation(data.sequence)
   else
     model:SetAnimation(0)